requirements.txt
ultralytics>=8.0.0
numpy>=1.24.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
PyYAML>=6.0.1
pytest>=7.0.0
pre-commit>=3.3.3
